Working on redesigning Cell Logic. currently I have implemented (but not tested) the following:
Level1.txt - contains data for loading a level.
   the first row of data is the mapHeight
   the second row of data is the mapwidth
   the rest of the data is the map of the level.Level1
ICell.cs - contains the definition of what a cell is.
IAnimateColorCell.cs - contains the definitions to allow a cell to render a material and Animate
ColorCell.cs - inherits the Cell logic and additionally handles the animations and material logic.
LevelLoader.cs - Loads a level from File using a set of information provided.
   the level loader uses a Dictionary to map characters to different cellGeneration functions.